requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.4
django-jazzmin==3.0.0
djangorestframework==3.15.1
drf-spectacular==0.27.2
dynaconf==3.2.5
hvac==2.2.0
idna==3.7
inflection==0.5.1
install==1.3.5
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
knox==0.1.14
loguru==0.7.2
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.0
shortuuid==1.0.13
sqlparse==0.5.0
uritemplate==4.1.1
urllib3==2.2.1
validators==0.28.1
